aws/requirements.txt
boto3>=1.26.0
requests>=2.28.0
Pillow>=9.0.0 